Fixed formatting issues, modularized components
