+++ src/components/header/CurrencyToggle/index.js
@@ -1,0 +1,1 @@
+export { default } from "./CurrencyToggle";
