+++ src/components/header/QRScanner/QRScanner.styles.js
