+++ src/components/header/QRScanner/index.js
@@ -1,0 +1,1 @@
+export { default } from "./QRScanner";
