+++ src/components/header/Searchbar/Searchbar.js
@@ -1,5 +1,5 @@
 import React from "react";
-import {Wrapper, Icon, Input } from "./Searchbar.styles"
+import { Wrapper, Icon, Input } from "./Searchbar.styles"
 
 export default class Searchbar extends React.Component {
     render() {
