+++ src/components/header/Searchbar/index.js
@@ -1,0 +1,1 @@
+export { default } from "./Searchbar";
