+++ src/components/index.js
@@ -1,0 +1,10 @@
+//Header components
+export { default as CurrencyToggle } from './header/CurrencyToggle';
+
+export { default as Header } from './header/Header';
+
+export { default as Infographic } from './header/Infographic';
+
+export { default as QRScanner } from './header/QRScanner';
+
+export { default as Searchbar } from './header/Searchbar';
