+++ src/pages/Coins/Coins.styles.js
@@ -1,3 +1,3 @@
-import styled from "styled-components"
+import styled from "styled-components";
 
 export const Wrapper = styled.div``;
