+++ src/pages/Portfolio/Portfolio.js
@@ -2,7 +2,6 @@ import React from "react"
 import {Wrapper} from "./Portfolio.styles"
 
 export default class Porfolio extends React.Component {
-
     render() {
         return (
             <Wrapper>
