+++ src/pages/Portfolio/index.js
@@ -1,1 +1,1 @@
-export {default} from "./Portfolio"
+export { default } from "./Portfolio";
