+++ src/pages/index.js
@@ -1,3 +1,3 @@
-export {default as Coins} from './Coins'
+export { default as Coins } from './Coins'
 
-export {default as Portfolio} from "./Portfolio"
+export { default as Portfolio } from "./Portfolio"
